typescript all uiParts
